pre xcode 6.3 beta

+++ stash/stash/GCMStoreCreator.swift
@@ -47,8 +47,8 @@ extension GCMStore {
             newStore = NSEntityDescription.insertNewObjectForEntityForName(GCMStoreClassNameKey, inManagedObjectContext: context) as GCMStore
             
             newStore.ciphertext         = resultBundle.cipherData!
-            newStore.scryptIterations   = Int64(i)
-            newStore.scryptMemoryFactor = Int64(N)
+            newStore.scryptIterations   = i
+            newStore.scryptMemoryFactor = N
             newStore.scryptSalt         = salt!
             newStore.verificationTag    = resultBundle.tag!
             newStore.nonce              = nonce!
